perf:Create a job that supports adding node parameters
