+++ apps/ops/models/job.py
@@ -29,6 +29,7 @@ from ops.ansible.exception import CommandInBlackListException
 from ops.mixin import PeriodTaskModelMixin
 from ops.variables import *
 from ops.const import Types, RunasPolicies, JobStatus, JobModules
+from ops.utils import merge_nodes_and_assets
 from orgs.mixins.models import JMSOrgBaseModel
 from perms.models import AssetPermission
 from perms.utils import UserPermAssetUtil
@@ -50,11 +51,13 @@ def get_parent_keys(key, include_self=True):
 class JMSPermedInventory(JMSInventory):
     def __init__(self,
                  assets,
+                 nodes,
                  account_policy='privileged_first',
                  account_prefer='root,Administrator',
                  module=None,
                  host_callback=None,
                  user=None):
+        assets = merge_nodes_and_assets(list(nodes), list(assets), user)
         super().__init__(assets, account_policy, account_prefer, host_callback, exclude_localhost=True)
         self.user = user
         self.module = module
@@ -149,7 +152,8 @@ class Job(JMSOrgBaseModel, PeriodTaskModelMixin):
     playbook = models.ForeignKey('ops.Playbook', verbose_name=_("Playbook"), null=True, on_delete=models.SET_NULL)
     type = models.CharField(max_length=128, choices=Types.choices, default=Types.adhoc, verbose_name=_("Type"))
     creator = models.ForeignKey('users.User', verbose_name=_("Creator"), on_delete=models.SET_NULL, null=True)
-    assets = models.ManyToManyField('assets.Asset', verbose_name=_("Assets"))
+    assets = models.ManyToManyField('assets.Asset', blank=True, verbose_name=_("Assets"))
+    nodes = models.ManyToManyField('assets.Node', blank=True, verbose_name=_("Node"))
     use_parameter_define = models.BooleanField(default=False, verbose_name=(_('Use Parameter Define')))
     parameters_define = models.JSONField(default=dict, verbose_name=_('Parameters define'))
     runas = models.CharField(max_length=128, default='root', verbose_name=_('Run as'))
@@ -203,7 +207,7 @@ class Job(JMSOrgBaseModel, PeriodTaskModelMixin):
 
     @property
     def inventory(self):
-        return JMSPermedInventory(self.assets.all(),
+        return JMSPermedInventory(self.assets.all(), self.nodes.all(),
                                   self.runas_policy, self.runas,
                                   user=self.creator, module=self.module)
 
@@ -220,7 +224,7 @@ class Job(JMSOrgBaseModel, PeriodTaskModelMixin):
 
     class Meta:
         verbose_name = _("Job")
-        unique_together = [('name', 'org_id', 'creator')]
+        unique_together = [('name', 'org_id', 'creator', 'type')]
         ordering = ['date_created']
