+++ apps/ops/utils.py
@@ -5,6 +5,9 @@ from django.conf import settings
 
 from common.utils import get_logger, make_dirs
 from jumpserver.const import PROJECT_DIR
+from perms.models import PermNode
+from perms.utils import UserPermAssetUtil
+from assets.models import Asset, Node
 
 logger = get_logger(__file__)
 
@@ -29,3 +32,19 @@ def get_ansible_log_verbosity(verbosity=0):
         return 1
     return verbosity
 
+
+def merge_nodes_and_assets(nodes, assets, user):
+    if not nodes:
+        return assets
+    perm_util = UserPermAssetUtil(user=user)
+    for node_id in nodes:
+        if isinstance(node_id, Node):
+            node_id = node_id.id
+        if node_id == PermNode.FAVORITE_NODE_KEY:
+            node_assets = perm_util.get_favorite_assets()
+        elif node_id == PermNode.UNGROUPED_NODE_KEY:
+            node_assets = perm_util.get_ungroup_assets()
+        else:
+            _, node_assets = perm_util.get_node_all_assets(node_id)
+        assets.extend(node_assets.exclude(id__in=[asset.id for asset in assets]))
+    return assets
